refactor(daemon): extract ops into daemon-ops module
