+++ src/server-daemon.ts
@@ -1,43 +1,27 @@
 import { closeSync, openSync } from "node:fs";
-import { mkdir, readdir, readFile, rm, stat, writeFile } from "node:fs/promises";
+import { mkdir, readdir, rm, writeFile } from "node:fs/promises";
 import { join } from "node:path";
-import { z } from "zod";
-import { type IsoDateTimeString, isoDateTimeSchema } from "./datetime";
+import {
+  clearStaleStartupLock,
+  daemonsDir,
+  isProcessAlive,
+  readServerLock,
+  type StartupLock,
+  serverLockPath,
+  serverLogPath,
+  startupLockPath,
+  writeServerLock,
+} from "./daemon-ops";
 import { PRIVATE_FILE_MODE } from "./file-ops";
-import { resolveHomeDir } from "./home-dir";
 import { t } from "./i18n";
 import { PROTOCOL_VERSION } from "./protocol";
 
-// 6        7
-// \_(ᴗ _ᴗ)_/
-const DEFAULT_PORT = 6767;
 const SERVER_START_TIMEOUT_MS = 10_000;
 const HEALTHCHECK_TIMEOUT_MS = 1_200;
 const STARTUP_LOCK_MAX_AGE_MS = 30_000;
 
-type ServerLock = {
-  pid: number;
-  port: number;
-  startedAt: IsoDateTimeString;
-};
-
-type StartupLock = {
-  pid: number;
-  port: number;
-  startedAt: IsoDateTimeString;
-};
-
-const serverLockSchema = z.object({
-  pid: z.number().int().positive(),
-  port: z.number().int().positive(),
-  startedAt: isoDateTimeSchema,
-});
-
-const startupLockSchema = z.object({
-  pid: z.number().int().positive(),
-  port: z.number().int().positive(),
-  startedAt: isoDateTimeSchema,
-});
+// Re-export for external consumers (cli-daemon.ts, etc.)
+export { serverLogPath } from "./daemon-ops";
 
 type EnsureLocalServerInput = {
   port: number;
@@ -66,61 +50,6 @@ type StopResult = {
 
 export function apiUrlForPort(port: number): string {
   return `http://127.0.0.1:${port}`;
-}
-
-function daemonsDir(homeDir = resolveHomeDir()): string {
-  return join(homeDir, ".acolyte", "daemons");
-}
-
-function daemonFileName(port: number, suffix: string): string {
-  return port === DEFAULT_PORT ? `server${suffix}` : `${port}${suffix}`;
-}
-
-function serverLockPath(port: number, homeDir = resolveHomeDir()): string {
-  return join(daemonsDir(homeDir), daemonFileName(port, ".lock"));
-}
-
-function startupLockPath(port: number, homeDir = resolveHomeDir()): string {
-  return join(daemonsDir(homeDir), daemonFileName(port, ".start.lock"));
-}
-
-export function serverLogPath(port: number, homeDir = resolveHomeDir()): string {
-  return join(daemonsDir(homeDir), daemonFileName(port, ".log"));
-}
-
-function parseServerLock(raw: string): ServerLock | null {
-  try {
-    const value = JSON.parse(raw);
-    const parsed = serverLockSchema.safeParse(value);
-    if (!parsed.success) return null;
-    return parsed.data;
-  } catch {
-    return null;
-  }
-}
-
-async function readServerLock(path: string): Promise<ServerLock | null> {
-  try {
-    const raw = await readFile(path, "utf8");
-    return parseServerLock(raw);
-  } catch {
-    return null;
-  }
-}
-
-async function writeServerLock(path: string, lock: ServerLock): Promise<void> {
-  await mkdir(join(path, ".."), { recursive: true });
-  await writeFile(path, JSON.stringify(lock), { encoding: "utf8", mode: PRIVATE_FILE_MODE });
-}
-
-function isProcessAlive(pid: number): boolean {
-  if (!Number.isInteger(pid) || pid <= 0) return false;
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch {
-    return false;
-  }
 }
 
 async function isServerHealthy(apiUrl: string, apiKey?: string, timeoutMs = HEALTHCHECK_TIMEOUT_MS): Promise<boolean> {
@@ -210,45 +139,6 @@ async function releaseStartupLock(path: string): Promise<void> {
   await rm(path, { force: true });
 }
 
-async function readStartupLock(path: string): Promise<StartupLock | number | null> {
-  try {
-    const raw = (await readFile(path, "utf8")).trim();
-    const parsedPid = Number(raw);
-    if (Number.isInteger(parsedPid) && parsedPid > 0) return parsedPid;
-    const parsedJson = startupLockSchema.safeParse(JSON.parse(raw));
-    if (parsedJson.success) return parsedJson.data;
-  } catch {
-    return null;
-  }
-  return null;
-}
-
-async function startupLockAgeMs(path: string, lock: StartupLock | number): Promise<number | null> {
-  if (typeof lock !== "number") {
-    const startedAt = Date.parse(lock.startedAt);
-    if (Number.isFinite(startedAt)) return Math.max(0, Date.now() - startedAt);
-  }
-  try {
-    const file = await stat(path);
-    return Math.max(0, Date.now() - file.mtimeMs);
-  } catch {
-    return null;
-  }
-}
-
-async function clearStaleStartupLock(path: string, maxAgeMs?: number): Promise<boolean> {
-  const lock = await readStartupLock(path);
-  const ownerPid = typeof lock === "number" ? lock : (lock?.pid ?? null);
-  if (lock !== null && ownerPid && isProcessAlive(ownerPid)) {
-    if (maxAgeMs === undefined) return false;
-    const ageMs = await startupLockAgeMs(path, lock);
-    if (ageMs === null || ageMs < maxAgeMs) return false;
-  }
-  if (lock === null && (await Bun.file(path).exists()) === false) return false;
-  await rm(path, { force: true });
-  return true;
-}
-
 async function waitForHealthyServerOrStaleStartupLock(
   apiUrl: string,
   apiKey: string | undefined,
@@ -265,9 +155,14 @@ async function waitForHealthyServerOrStaleStartupLock(
   throw new Error(t("cli.server.start_timeout", { url: apiUrl }));
 }
 
-async function cleanupLegacyLocks(homeDir = resolveHomeDir()): Promise<void> {
-  await rm(join(homeDir, ".acolyte", "server.lock"), { force: true });
-  await rm(join(homeDir, ".acolyte", "server.start.lock"), { force: true });
+// 6        7
+// \_(ᴗ _ᴗ)_/
+const DEFAULT_PORT = 6767;
+
+async function cleanupLegacyLocks(homeDir?: string): Promise<void> {
+  const home = homeDir ?? "";
+  await rm(join(home, ".acolyte", "server.lock"), { force: true });
+  await rm(join(home, ".acolyte", "server.start.lock"), { force: true });
   await rm(join(daemonsDir(homeDir), `${DEFAULT_PORT}.lock`), { force: true });
   await rm(join(daemonsDir(homeDir), `${DEFAULT_PORT}.start.lock`), { force: true });
   await rm(join(daemonsDir(homeDir), `${DEFAULT_PORT}.log`), { force: true });
@@ -456,13 +351,3 @@ export async function listRunningDaemons(input?: {
   }
   return daemons.sort((a, b) => a.port - b.port);
 }
-
-export const serverDaemonInternals = {
-  daemonsDir,
-  serverLockPath,
-  startupLockPath,
-  serverLogPath,
-  parseServerLock,
-  isProcessAlive,
-  clearStaleStartupLock,
-};
